requirements.txt
langchain>=0.1.0
langchain_community>=0.0.10
langchain-openai>=0.0.5
supabase>=2.0.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.8.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
numpy>=1.24.0
pypdf>=3.15.0
unstructured>=0.10.0
pillow>=10.0.0
pytesseract>=0.3.10
python-multipart>=0.0.6
tiktoken>=0.5.0
faiss-cpu>=1.7.4
requests>=2.31.0
uvicorn
fastapi